pyproject.toml
[tool.poetry]
name = "python-restful-booker"
version = "0.1.0"
description = ""
authors = ["Anna Voitiuk"]
package-mode = false
readme = "README.md"

[tool.poetry.dependencies]
jsonschema = "^4.25.1"
pydantic = "^2.12.5"
flake8 = "^7.3.0"
faker = "^38.2.0"
pyyaml = "^6.0.3"
pydantic-settings = "^2.12.0"
pre-commit = "^3.8.0"
python = ">=3.10,<3.13.5"
colorlog = "^6.10.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
ruff = "^0.14.10"
pytest = "^9.0.2"
assertpy = "^1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
builtins = ["_"]
line-length = 120
extend-exclude = [
    "tests",
    "src/helpers"
]

[tool.ruff.lint]
extend-select = [
    "E501", # Add the 'line-too-long' rule to the enforced rule set
    "D", # Add pydocstyle
    "I" # Add Isort
]

[tool.ruff.lint.pydocstyle]
convention = "google"


[tool.pytest.ini_options]
pythonpath = ["."]
markers = [
    "smoke: used to check the main functionality and accept the build",
    "regression: used to execute the regression test cycles"
    ]